+++ sdk/lib/_internal/js_dev_runtime/patch/core_patch.dart
@@ -357,15 +357,6 @@ class Error {
   }
 }
 
-@patch
-class FallThroughError {
-  @patch
-  FallThroughError._create(String url, int line);
-
-  @patch
-  String toString() => super.toString();
-}
-
 // Patch for DateTime implementation.
 @patch
 class DateTime {