+++ tests/language/inference_update_1/write_capture_deferral_enabled_test.dart
@@ -8,8 +8,6 @@
 // naturally out of the fact that closures are analyzed last (so that their
 // types can depend on the types of other arguments).
 
-// SharedOptions=--enable-experiment=inference-update-1
-
 import '../static_type_helper.dart';
 
 withUnnamedArguments(int? i, void Function(void Function(), Object?) f) {